requirements.txt
aioredis==1.3.1
alabaster==0.7.12
amqp==5.1.0
anyascii==0.3.0
anyio==2.2.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
arrow==0.13.1
asgiref==3.5.0
asn1crypto==1.4.0
astroid==2.6.6
astropy==4.3.1
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
ats-utilities==2.2.4
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
autopep8==1.5.7
autoscraper==1.1.12
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
binaryornot==0.4.4
bitarray==2.3.0
bkcharts==0.2
black==19.10b0
bleach==4.0.0
bokeh==2.4.1
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
bs4==0.0.1
CacheControl==0.12.10
cached-property==1.5.2
cachetools==5.0.0
celery==5.2.3
certifi==2021.10.8
cffi==1.14.6
channels==3.0.4
channels-redis==3.4.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.4
comtypes==1.1.10
configparser==5.2.0
constantly==15.1.0
construct==2.5.3
contextlib2==0.6.0.post1
cookiecutter==1.7.2
cryptography==3.4.8
cycler==0.10.0
Cython==0.29.24
cytoolz==0.11.0
daal4py==2021.3.0
daphne==3.0.2
dask==2021.10.0
debugpy==1.4.1
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dist-py-module==2.3.8
distlib==0.3.4
distributed==2021.10.0
Django==3.2.12
django-allauth==0.49.0
django-core==1.4.1
django-countries==7.3.2
django-crispy-forms==1.14.0
django-crontab==0.7.1
django-debug-toolbar==3.2.4
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
django-extensions==3.1.5
django-extra-fields==3.0.2
django-filter==21.1
django-haystack==3.1.1
django-jet==1.0.8
django-js-asset==2.0.0
django-markdown-deux==1.0.5
django-model-utils==4.2.0
django-modelcluster==5.2
django-mptt==0.13.4
django-nine==0.2.5
django-notifications-hq==1.7.0
django-pagedown==2.2.1
django-phonenumber-field==6.1.0
django-randompinfield==0.4.7
django-rest-auth==0.9.5
django-rest-knox==4.2.0
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.17.1
draftjs-exporter==2.1.7
drf-haystack==1.8.11
elasticsearch==7.17.1
elasticsearch-dsl==7.4.0
entrypoints==0.3
et-xmlfile==1.1.0
fastcache==1.1.0
fcm-django==1.0.8
filelock==3.3.1
firebase-admin==5.2.0
flake8==3.9.2
Flask==1.1.2
fonttools==4.25.0
fsspec==2021.10.1
future==0.18.2
gevent==21.8.0
glob2==0.7
google-api-core==2.7.1
google-api-python-client==2.40.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.55.0
googletrans==3.0.0
greenlet==1.1.1
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.9.0
h2==3.2.0
h5py==3.2.1
haystack==0.42
HeapDict==1.0.1
hiredis==2.0.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imagecodecs==2021.8.26
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.8.1
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
intervaltree==3.1.0
ipykernel==6.4.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
jsonfield==3.1.0
jsonschema==4.4.0
keyring==23.1.0
kiwisolver==1.3.1
kombu==5.2.4
l18n==2021.3
lazy-object-proxy==1.6.0
libarchive-c==2.9
llvmlite==0.37.0
locket==0.2.1
lxml==4.6.3
Markdown==3.3.6
markdown2==2.4.2
MarkupSafe==1.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mccabe==0.6.1
menuinst==1.4.18
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
more-itertools==8.10.0
mpmath==1.2.1
msgpack==1.0.2
multipledispatch==0.6.0
munkres==1.1.4
mypy-extensions==0.4.3
mysqlclient==2.1.0
navigator-updater==0.2.1
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.6.5
nose==1.3.7
notebook==6.4.5
numba==0.54.1
numexpr==2.7.3
numpy==1.20.3
numpydoc==1.1.0
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.8.2
partd==1.2.0
path==16.0.0
pathlib2==2.3.6
pathspec==0.7.0
patsy==0.5.2
pefile==2021.9.3
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.44
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
pip-review==1.1.1
pipdeptree==2.2.1
pipreqs==0.4.11
pkginfo==1.7.1
platformdirs==2.5.1
pluggy==0.13.1
ply==3.11
poyo==0.5.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
proto-plus==1.20.3
protobuf==3.19.4
proxycrawl==3.2.1
psutil==5.8.0
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.21
pycryptodome==3.14.1
pycurl==7.44.1
pydocstyle==6.1.1
pyerfa==2.0.0
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==1.7.1
pylint==2.9.6
pyls-spyder==0.4.0
PyNaCl==1.4.0
pyodbc==4.0.0-unsupported
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyreadline==2.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.2
python-decouple==3.6
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-memcached==1.59
python-ptrace==0.9.8
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
PyWavelets==1.1.1
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.2
qtconsole==5.1.1
QtPy==1.10.0
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rope==0.19.0
rsa==4.8
Rtree==0.9.7
ruamel-yaml-conda==0.15.100
scikit-image==0.18.3
scikit-learn==0.24.2
scikit-learn-intelex==2021.20210714.120553
scipy==1.7.1
seaborn==0.11.2
selenium==4.1.3
Send2Trash==1.8.0
serpy==0.3.1
service-identity==21.1.0
setuptools==59.6.0
simplegeneric==0.8.1
singledispatch==3.7.0
sip==4.19.13
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.22
sqlparse==0.4.2
statsmodels==0.12.2
swapper==1.3.0
sympy==1.9
tables==3.6.1
tablib==3.2.0
TBB==0.2
tblib==1.7.0
telepath==0.2
terminado==0.9.4
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
twilio==7.7.1
Twisted==22.2.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
tzdata==2021.5
ujson==4.0.2
unicodecsv==0.14.1
Unidecode==1.2.0
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.13.3
wagtail==2.16.1
watchdog==2.1.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wheel==0.37.0
whichcraft==0.6.1
widgetsnbextension==3.5.1
Willow==1.4.1
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.1
xlwings==0.24.9
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.31.0
yarg==0.1.9
zict==2.0.0
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0
